backend/requirements.txt
Django==5.1.4
djangorestframework
django-cors-headers
Pillow
python-dotenv
google-generativeai
gunicorn
